refactor: move ecoding_tools from industry to dataspace
