+++ src/tractusx_sdk/dataspace/tools/__init__.py
@@ -29,3 +29,4 @@ __license__ = "Apache License, Version 2.0"
 
 from .http_tools import HttpTools
 from .operators import op
+from .encoding_tools import encode_as_base64_url_safe, decode_base64_url_safe
